chore: move Next.js folders to root for proper framework detection
