+++ components/ProjectGallery.tsx
