+++ pages/AIPage.tsx
