+++ pages/ProjectDetailPage.tsx
