--- src/components/Layout.tsx
@@ -1,26 +1,0 @@
-import { ReactNode, useEffect } from 'react';
-import { usePathname } from 'next/navigation';
-import Header from './Header';
-import Footer from './Footer';
-
-interface LayoutProps {
-  children: ReactNode;
-}
-
-const Layout = ({ children }: LayoutProps) => {
-  const pathname = usePathname();
-
-  useEffect(() => {
-    window.scrollTo(0, 0);
-  }, [pathname]);
-
-  return (
-    <div className="flex flex-col min-h-screen">
-      <Header />
-      <main className="flex-grow">{children}</main>
-      <Footer />
-    </div>
-  );
-};
-
-export default Layout;
