--- src/context/ProjectContext.ts
@@ -1,9 +1,0 @@
-import { createContext } from 'react';
-import { Project } from '../types/Project';
-
-export interface ProjectContextType {
-  projects: Project[];
-}
-
-export const ProjectContext = createContext<ProjectContextType | undefined>(undefined);
-
