--- src/hooks/useProjectContext.ts
@@ -1,11 +1,0 @@
-import { useContext } from 'react';
-import { ProjectContext } from '../context/ProjectContext';
-
-export const useProjectContext = () => {
-  const context = useContext(ProjectContext);
-  if (context === undefined) {
-    throw new Error('useProjectContext must be used within a ProjectProvider');
-  }
-  return context;
-};
-
